Move shared interface modules into ui package
