+++ zapzap/ui/__init__.py
@@ -1,0 +1,1 @@
+"""Shared UI modules for ZapZap."""
