+++ zapzap/ui/main_window/client_side_rendering_view.py
