+++ zapzap/views/settings_components/settings_section.py
@@ -1,6 +1,6 @@
 from PyQt6.QtWidgets import QVBoxLayout, QWidget
 
-from zapzap.views.components import Label
+from zapzap.ui.components import Label
 
 
 class SettingsSection(QWidget):
